Refactor refactored code after rebase
